🚚(backend) split views into multiple files
We are going to develop new view. The views.py module starts to be too long. So
we decide to split this file into two files : debug and redirect views.
